+++ data.json
@@ -15,6 +15,22 @@
       "price": 20,
       "status": "declined",
       "user_id": 1
+    },
+    {
+      "id": 9,
+      "food": [
+        "meat",
+        "fish",
+        "plantain"
+      ],
+      "quantity": [
+        "1",
+        "3",
+        "4"
+      ],
+      "price": 429,
+      "status": "declined",
+      "user_id": 2
     }
   ],
   "foodList": [
